+++ eq-author/.storybook/main.js
@@ -5,5 +5,6 @@ module.exports = {
     "@storybook/addon-essentials",
     "@storybook/addon-controls",
     "@storybook/addon-a11y",
+    "storybook-dark-mode",
   ],
 };
