+++ eq-author/.storybook/preview.js
@@ -1,6 +1,10 @@
 import React from "react";
 import { Router, Route } from "react-router-dom";
 import { createMemoryHistory } from "history";
+import { addDecorator } from "@storybook/react";
+import { withConsole } from "@storybook/addon-console";
+
+addDecorator((storyFn, context) => withConsole()(storyFn)(context));
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
