+++ eq-author/package.json
@@ -24,11 +24,12 @@
     "@babel/preset-env": "latest",
     "@babel/preset-react": "latest",
     "@hot-loader/react-dom": "latest",
-    "@storybook/addon-a11y": "^6.0.28",
-    "@storybook/addon-actions": "^6.0.21",
-    "@storybook/addon-essentials": "^6.0.18",
-    "@storybook/addon-links": "^6.1.18",
-    "@storybook/react": "^6.0.18",
+    "@storybook/addon-a11y": "^6.1.21",
+    "@storybook/addon-actions": "^6.1.21",
+    "@storybook/addon-console": "^1.2.3",
+    "@storybook/addon-essentials": "^6.1.21",
+    "@storybook/addon-links": "^6.1.21",
+    "@storybook/react": "^6.1.21",
     "@testing-library/jest-dom": "latest",
     "@testing-library/react": "latest",
     "@testing-library/user-event": "^12.2.2",
@@ -75,6 +76,7 @@
     "react-svg-loader": "latest",
     "redux-mock-store": "latest",
     "sass-loader": "latest",
+    "storybook-dark-mode": "^1.0.7",
     "style-loader": "latest",
     "stylelint": "latest",
     "stylelint-config-standard": "^17.0.0",
