+++ eq-author/src/App/Comments/graphql/createNewComment.graphql
