+++ eq-author/src/App/Comments/graphql/deleteComment.graphql
