+++ eq-author/src/App/Comments/graphql/deleteReply.graphql
