+++ eq-author/src/components/Comments/CommentEditor/index.js
@@ -1,0 +1,147 @@
+import React, { useState } from "react";
+import styled from "styled-components";
+import PropTypes from "prop-types";
+
+import { colors, focusStyle } from "constants/theme";
+
+import Button from "components/buttons/Button";
+import TextareaAutosize from "react-textarea-autosize";
+
+const Wrapper = styled.div`
+  margin-bottom: 1em;
+`;
+
+const TextAreaFixedSize = styled.textarea`
+  height: 94px;
+  width: 100%;
+  border: thin solid ${colors.grey};
+  resize: none;
+  font-size: 1em;
+  font-family: inherit;
+  padding: 0.5em;
+  margin-bottom: 0.5em;
+
+  &:focus {
+    ${focusStyle}
+    outline: none;
+  }
+`;
+
+const TextAreaGrowable = styled(TextareaAutosize)`
+  width: 100%;
+  border: thin solid ${colors.grey};
+  resize: none;
+  font-size: 1em;
+  font-family: inherit;
+  padding: 0.5em;
+  margin-bottom: 0.5em;
+
+  &:focus {
+    ${focusStyle}
+    outline: none;
+  }
+`;
+
+const ButtonGroup = styled.div`
+  display: flex;
+
+  button {
+    margin-right: 0.5em;
+  }
+`;
+
+const CommentEditor = ({
+  confirmText = "Add",
+  cancelText = "Cancel",
+  variant = "fixed",
+  initialValue = "",
+  showCancel = false,
+  onCancel,
+  onConfirm,
+  commentId,
+  replyId,
+  className,
+}) => {
+  const [commentText, updateCommentText] = useState(initialValue);
+
+  return (
+    <Wrapper className={className}>
+      {variant === "fixed" && (
+        <TextAreaFixedSize
+          value={commentText}
+          onChange={({ target: { value } }) => updateCommentText(value)}
+        />
+      )}
+      {variant === "growable" && (
+        <TextAreaGrowable
+          value={commentText}
+          onChange={({ target: { value } }) => updateCommentText(value)}
+        />
+      )}
+      <ButtonGroup>
+        <Button
+          type="submit"
+          variant="greyed"
+          small-medium
+          disabled={commentText.length === 0}
+          onClick={() => {
+            onConfirm(commentText, commentId, replyId);
+            updateCommentText("");
+          }}
+        >
+          {confirmText}
+        </Button>
+        {showCancel && (
+          <Button variant="greyed" small-medium onClick={onCancel}>
+            {cancelText}
+          </Button>
+        )}
+      </ButtonGroup>
+    </Wrapper>
+  );
+};
+
+CommentEditor.propTypes = {
+  /**
+   * The text to show on the confirm button.
+   */
+  confirmText: PropTypes.string,
+  /**
+   * The text to show on the cancel button.
+   */
+  cancelText: PropTypes.string,
+  /**
+   * The value to show in the input box when it is first rendered.
+   */
+  initialValue: PropTypes.string,
+  /**
+   * The style of the input box.
+   */
+  variant: PropTypes.oneOf(["fixed", "growable"]),
+  /**
+   * When true, the 'Cancel' button is shown.
+   */
+  showCancel: PropTypes.bool,
+  /**
+   * A function that runs when the cancel button is pressed. Required if 'showCancel' is true.
+   */
+  onCancel: PropTypes.func,
+  /**
+   * A function that runs when the 'Add' button is pressed.
+   */
+  onConfirm: PropTypes.func.isRequired,
+  /**
+   * If given, it is made available to the 'onConfirm' function. Useful when adding replies to a root comment.
+   */
+  commentId: PropTypes.string,
+  /**
+   * If given, it is made available to the 'onConfirm' function. Useful when adding replies to the root comment.
+   */
+  replyId: PropTypes.string,
+  /**
+   * Allows for CSS classes to be filtered down when using Styled-Components.
+   */
+  className: PropTypes.string,
+};
+
+export default CommentEditor;
