+++ eq-author/src/storybook/Patterns/Comments/PureComment.stories.js
@@ -1,0 +1,44 @@
+import React from "react";
+
+import {
+  Title,
+  Description,
+  Primary,
+  ArgsTable,
+  Stories,
+  PRIMARY_STORY,
+} from "@storybook/addon-docs/blocks";
+
+import PureComment from "components/Comments/PureComment";
+
+export default {
+  title: "Patterns/Comments/Pure Comment",
+  component: PureComment,
+  parameters: {
+    docs: {
+      page: () => (
+        <>
+          <Title />
+          <Description>
+            A basic comment that facilitates modifying and deleting.
+          </Description>
+          <Primary />
+          <ArgsTable story={PRIMARY_STORY} />
+          <Stories />
+        </>
+      ),
+    },
+  },
+};
+
+const PureCommentTemplate = (args) => <PureComment {...args} />;
+
+export const Default = PureCommentTemplate.bind({});
+Default.args = {
+  commentId: "1",
+  author: "Jane Doe",
+  datePosted: "Today at 15:43",
+  text: "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
+  onUpdateComment: () => alert("Comment updated"),
+  onDeleteComment: () => alert("Comment deleted"),
+};
